flowedit: Move content_extents to FlowCanvasState in window_state.rs (#2602)
Move content_extents from flow_canvas.rs to a private method on
FlowCanvasState since it's only called from auto_fit.
